Thin opensds feature

+++ pkg/api/controllers/volume.go
@@ -23,13 +23,13 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-
 	log "github.com/golang/glog"
 	"github.com/opensds/opensds/pkg/api/policy"
 	"github.com/opensds/opensds/pkg/api/util"
 	c "github.com/opensds/opensds/pkg/context"
 	"github.com/opensds/opensds/pkg/controller/client"
 	"github.com/opensds/opensds/pkg/db"
+	dock "github.com/opensds/opensds/pkg/dock/client"
 	"github.com/opensds/opensds/pkg/model"
 	pb "github.com/opensds/opensds/pkg/model/proto"
 	. "github.com/opensds/opensds/pkg/utils/config"
@@ -37,14 +37,15 @@ import (
 
 func NewVolumePortal() *VolumePortal {
 	return &VolumePortal{
-		CtrClient: client.NewClient(),
+		CtrClient:  client.NewClient(),
+		DockClient: dock.NewClient(),
 	}
 }
 
 type VolumePortal struct {
 	BasePortal
-
-	CtrClient client.Client
+	CtrClient  client.Client
+	DockClient dock.Client
 }
 
 func (v *VolumePortal) CreateVolume() {
@@ -69,16 +70,30 @@ func (v *VolumePortal) CreateVolume() {
 	if volume.ProfileId == "" {
 		log.Warning("Use default profile when user doesn't specify profile.")
 		prf, err = db.C.GetDefaultProfile(ctx)
+		if err != nil {
+			errMsg := fmt.Sprintf("get profile failed: %s", err.Error())
+			v.ErrorHandle(model.ErrorBadRequest, errMsg)
+			return
+		}
 		volume.ProfileId = prf.Id
 	} else {
 		prf, err = db.C.GetProfile(ctx, volume.ProfileId)
+		if err != nil {
+			errMsg := fmt.Sprintf("get profile failed: %s", err.Error())
+			v.ErrorHandle(model.ErrorBadRequest, errMsg)
+			return
+		}
 	}
-	if err != nil {
-		errMsg := fmt.Sprintf("get profile failed: %s", err.Error())
-		v.ErrorHandle(model.ErrorBadRequest, errMsg)
-		return
+	// To get pool details for Thin OpenSDS
+	var pools []*model.StoragePoolSpec
+	if CONF.OsdsApiServer.InstallType == "thin" {
+		pools, err = db.C.ListPools(c.NewAdminContext())
+		if err != nil {
+			log.Error("when listing pools: ", err)
+			return
+		}
+		volume.PoolId = pools[0].Id
 	}
-
 	// NOTE:It will create a volume entry into the database and initialize its status
 	// as "creating". It will not wait for the real volume creation to complete
 	// and will return result immediately.
@@ -88,7 +103,6 @@ func (v *VolumePortal) CreateVolume() {
 		v.ErrorHandle(model.ErrorBadRequest, errMsg)
 		return
 	}
-
 	// Marshal the result.
 	body, _ := json.Marshal(result)
 	v.SuccessHandle(StatusAccepted, body)
@@ -96,11 +110,19 @@ func (v *VolumePortal) CreateVolume() {
 	// NOTE:The real volume creation process.
 	// Volume creation request is sent to the Dock. Dock will update volume status to "available"
 	// after volume creation is completed.
-	if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.CreateVolumeOpts{
 		Id:               result.Id,
@@ -117,12 +139,21 @@ func (v *VolumePortal) CreateVolume() {
 		SnapshotFromCloud: result.SnapshotFromCloud,
 		Context:           ctx.ToJson(),
 	}
-	if _, err = v.CtrClient.CreateVolume(context.Background(), opt); err != nil {
-		log.Error("create volume failed in controller service:", err)
+
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if _, err = v.CtrClient.CreateVolume(context.Background(), opt); err != nil {
+			log.Error("create volume failed in controller service:", err)
+			return
+		}
+	} else {
+		opt.DriverName = CONF.OsdsDock.EnabledBackends[0]
+		opt.PoolName = pools[0].Name
+		if _, err := v.DockClient.CreateVolume(context.Background(), opt); err != nil {
+			log.Error("create volume failed in controller service:", err)
+			return
+		}
 		return
 	}
-
-	return
 }
 
 func (v *VolumePortal) ListVolumes() {
@@ -136,14 +167,12 @@ func (v *VolumePortal) ListVolumes() {
 		v.ErrorHandle(model.ErrorBadRequest, errMsg)
 		return
 	}
-
 	result, err := db.C.ListVolumesWithFilter(c.GetContext(v.Ctx), m)
 	if err != nil {
 		errMsg := fmt.Sprintf("list volumes failed: %s", err.Error())
 		v.ErrorHandle(model.ErrorInternalServer, errMsg)
 		return
 	}
-
 	// Marshal the result.
 	body, _ := json.Marshal(result)
 	v.SuccessHandle(StatusOK, body)
@@ -233,23 +262,64 @@ func (v *VolumePortal) ExtendVolume() {
 	// NOTE:The real volume extension process.
 	// Volume extension request is sent to the Dock. Dock will update volume status to "available"
 	// after volume extension is completed.
-	if err = v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	var pool *model.StoragePoolSpec
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		// To roll back size and status if INVALID
+		var rollBack = false
+		defer func() {
+			if rollBack {
+				db.UpdateVolumeStatus(ctx, db.C, result.Id, model.VolumeAvailable)
+				log.Info("extend volume failed due to invalid size")
+			}
+		}()
+
+		pool, err = db.C.GetPool(ctx, result.PoolId)
+		if nil != err {
+			log.Error("get pool failed in extend volume method: ", err.Error())
+			rollBack = true
+			return
+		}
+
+		var newSize = extendRequestBody.NewSize
+		if pool.FreeCapacity <= (newSize - result.Size) {
+			log.Info("pool free capacity is less than volume size requested")
+			rollBack = true
+			return
+		}
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.ExtendVolumeOpts{
 		Id:       id,
 		Size:     extendRequestBody.NewSize,
 		Metadata: result.Metadata,
 		Context:  ctx.ToJson(),
-	}
-	if _, err = v.CtrClient.ExtendVolume(context.Background(), opt); err != nil {
-		log.Error("extend volume failed in controller service:", err)
-		return
+		PoolId:   result.PoolId,
 	}
 
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if _, err = v.CtrClient.ExtendVolume(context.Background(), opt); err != nil {
+			log.Error("extend volume failed in controller service:", err)
+			return
+		}
+	} else {
+		opt.PoolName = pool.Name
+		opt.DriverName = CONF.OsdsDock.EnabledBackends[0]
+		if _, err := v.DockClient.ExtendVolume(context.Background(), opt); err != nil {
+			log.Error("extend volume failed in controller service:", err)
+			return
+		}
+	}
 	return
 }
 
@@ -267,7 +337,6 @@ func (v *VolumePortal) DeleteVolume() {
 		v.ErrorHandle(model.ErrorNotFound, errMsg)
 		return
 	}
-
 	// If profileId or poolId of the volume doesn't exist, it would mean that
 	// the volume provisioning operation failed before the create method in
 	// storage driver was called, therefore the volume entry should be deleted
@@ -302,11 +371,20 @@ func (v *VolumePortal) DeleteVolume() {
 	// NOTE:The real volume deletion process.
 	// Volume deletion request is sent to the Dock. Dock will delete volume from driver
 	// and database or update volume status to "errorDeleting" if deletion from driver faild.
-	if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
+
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.DeleteVolumeOpts{
 		Id:        volume.Id,
@@ -316,24 +394,38 @@ func (v *VolumePortal) DeleteVolume() {
 		Context:   ctx.ToJson(),
 		Profile:   prf.ToJson(),
 	}
-	if _, err = v.CtrClient.DeleteVolume(context.Background(), opt); err != nil {
-		log.Error("delete volume failed in controller service:", err)
-		return
-	}
 
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if _, err = v.CtrClient.DeleteVolume(context.Background(), opt); err != nil {
+			log.Error("delete volume failed in controller service:", err)
+			return
+		}
+	} else {
+
+		opt.DriverName = CONF.OsdsDock.EnabledBackends[0]
+		if _, err := v.DockClient.DeleteVolume(context.Background(), opt); err != nil {
+			log.Error("delete volume failed in controller service:", err)
+			return
+		}
+		if err = db.C.DeleteVolume(ctx, opt.GetId()); err != nil {
+			return
+		}
+	}
 	return
 }
 
 func NewVolumeAttachmentPortal() *VolumeAttachmentPortal {
 	return &VolumeAttachmentPortal{
-		CtrClient: client.NewClient(),
+		CtrClient:  client.NewClient(),
+		DockClient: dock.NewClient(),
 	}
 }
 
 type VolumeAttachmentPortal struct {
 	BasePortal
 
-	CtrClient client.Client
+	CtrClient  client.Client
+	DockClient dock.Client
 }
 
 func (v *VolumeAttachmentPortal) CreateVolumeAttachment() {
@@ -368,11 +460,19 @@ func (v *VolumeAttachmentPortal) CreateVolumeAttachment() {
 	// NOTE:The real volume attachment creation process.
 	// Volume attachment creation request is sent to the Dock. Dock will update volume attachment status to "available"
 	// after volume attachment creation is completed.
-	if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.CreateVolumeAttachmentOpts{
 		Id:       result.Id,
@@ -387,9 +487,37 @@ func (v *VolumeAttachmentPortal) CreateVolumeAttachment() {
 		Metadata: result.Metadata,
 		Context:  ctx.ToJson(),
 	}
-	if _, err = v.CtrClient.CreateVolumeAttachment(context.Background(), opt); err != nil {
-		log.Error("create volume attachment failed in controller service:", err)
-		return
+
+	if CONF.OsdsApiServer.InstallType == "thin" {
+		vol, err := db.C.GetVolume(ctx, result.VolumeId)
+		if err != nil {
+			log.Error("create volume attachment failed in controller service")
+			return
+		}
+
+		pol, err := db.C.GetPool(ctx, vol.PoolId)
+		if err != nil {
+			log.Error("get pool failed in create volume attachment method: ", err)
+			db.UpdateVolumeAttachmentStatus(ctx, db.C, opt.Id, model.VolumeAttachError)
+			return
+		}
+		var protocol = pol.Extras.IOConnectivity.AccessProtocol
+		if protocol == "" {
+			// Default protocol is iscsi
+			protocol = "iscsi"
+		}
+		opt.AccessProtocol = protocol
+		opt.Metadata = vol.Metadata
+		opt.DriverName = CONF.OsdsDock.EnabledBackends[0]
+		if _, err := v.DockClient.CreateVolumeAttachment(context.Background(), opt); err != nil {
+			log.Error("create volume attachment failed in controller service:", err)
+			return
+		}
+	} else {
+		if _, err = v.CtrClient.CreateVolumeAttachment(context.Background(), opt); err != nil {
+			log.Error("create volume attachment failed in controller service:", err)
+			return
+		}
 	}
 
 	return
@@ -492,11 +620,19 @@ func (v *VolumeAttachmentPortal) DeleteVolumeAttachment() {
 	// NOTE:The real volume attachment deletion process.
 	// Volume attachment deletion request is sent to the Dock. Dock will delete volume attachment from database
 	// or update its status to "errorDeleting" if volume connection termination failed.
-	if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.DeleteVolumeAttachmentOpts{
 		Id:             attachment.Id,
@@ -512,9 +648,16 @@ func (v *VolumeAttachmentPortal) DeleteVolumeAttachment() {
 		Metadata: attachment.Metadata,
 		Context:  ctx.ToJson(),
 	}
-	if _, err = v.CtrClient.DeleteVolumeAttachment(context.Background(), opt); err != nil {
-		log.Error("delete volume attachment failed in controller service:", err)
-		return
+	if CONF.OsdsApiServer.InstallType == "thin" {
+		if _, err = v.DockClient.DeleteVolumeAttachment(context.Background(), opt); err != nil {
+			log.Error("delete volume attachment failed in controller service:", err)
+			return
+		}
+	} else {
+		if _, err = v.CtrClient.DeleteVolumeAttachment(context.Background(), opt); err != nil {
+			log.Error("delete volume attachment failed in controller service:", err)
+			return
+		}
 	}
 
 	return
@@ -522,14 +665,15 @@ func (v *VolumeAttachmentPortal) DeleteVolumeAttachment() {
 
 func NewVolumeSnapshotPortal() *VolumeSnapshotPortal {
 	return &VolumeSnapshotPortal{
-		CtrClient: client.NewClient(),
+		CtrClient:  client.NewClient(),
+		DockClient: dock.NewClient(),
 	}
 }
 
 type VolumeSnapshotPortal struct {
 	BasePortal
-
-	CtrClient client.Client
+	DockClient dock.Client
+	CtrClient  client.Client
 }
 
 func (v *VolumeSnapshotPortal) CreateVolumeSnapshot() {
@@ -557,6 +701,21 @@ func (v *VolumeSnapshotPortal) CreateVolumeSnapshot() {
 		return
 	}
 
+	if CONF.OsdsApiServer.InstallType == "thin" {
+		// Get snapshot profile
+		if result.ProfileId != "" {
+			profile, err := db.C.GetProfile(ctx, result.ProfileId)
+			if err != nil {
+				log.Error("when get profile resource: ", err)
+				db.UpdateVolumeSnapshotStatus(ctx, db.C, result.Id, model.VolumeSnapError)
+				return
+			}
+
+			if profile.SnapshotProperties.Topology.Bucket != "" {
+				result.Metadata["bucket"] = profile.SnapshotProperties.Topology.Bucket
+			}
+		}
+	}
 	// Marshal the result.
 	body, _ := json.Marshal(result)
 	v.SuccessHandle(StatusAccepted, body)
@@ -564,11 +723,19 @@ func (v *VolumeSnapshotPortal) CreateVolumeSnapshot() {
 	// NOTE:The real volume snapshot creation process.
 	// Volume snapshot creation request is sent to the Dock. Dock will update volume snapshot status to "available"
 	// after volume snapshot creation complete.
-	if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.CreateVolumeSnapshotOpts{
 		Id:          result.Id,
@@ -579,11 +746,26 @@ func (v *VolumeSnapshotPortal) CreateVolumeSnapshot() {
 		Metadata:    result.Metadata,
 		Context:     ctx.ToJson(),
 	}
-	if _, err = v.CtrClient.CreateVolumeSnapshot(context.Background(), opt); err != nil {
-		log.Error("create volume snapthot failed in controller service:", err)
-		return
-	}
 
+	if CONF.OsdsApiServer.InstallType == "thin" {
+		volume, err := db.C.GetVolume(ctx, result.VolumeId)
+		if err != nil {
+			log.Error("create volume snapshot failed in controller service")
+			return
+		}
+		opt.Metadata = volume.Metadata
+		opt.Size = volume.Size
+		opt.DriverName = CONF.OsdsDock.EnabledBackends[0]
+		if _, err = v.DockClient.CreateVolumeSnapshot(context.Background(), opt); err != nil {
+			log.Error("create volume snapshot failed in controller service:", err)
+			return
+		}
+	} else {
+		if _, err = v.CtrClient.CreateVolumeSnapshot(context.Background(), opt); err != nil {
+			log.Error("create volume snapshot failed in controller service:", err)
+			return
+		}
+	}
 	return
 }
 
@@ -689,11 +871,19 @@ func (v *VolumeSnapshotPortal) DeleteVolumeSnapshot() {
 	// NOTE:The real volume snapshot deletion process.
 	// Volume snapshot deletion request is sent to the Dock. Dock will delete volume snapshot from driver and
 	// database or update its status to "errorDeleting" if volume snapshot deletion from driver failed.
-	if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
-		log.Error("when connecting controller client:", err)
-		return
+	if CONF.OsdsApiServer.InstallType != "thin" {
+		if err := v.CtrClient.Connect(CONF.OsdsLet.ApiEndpoint); err != nil {
+			log.Error("when connecting controller client:", err)
+			return
+		}
+	} else {
+		if err := v.DockClient.Connect(CONF.OsdsDock.ApiEndpoint); err != nil {
+			log.Error("when connecting dock client:", err)
+			return
+		}
 	}
 	defer v.CtrClient.Close()
+	defer v.DockClient.Close()
 
 	opt := &pb.DeleteVolumeSnapshotOpts{
 		Id:       snapshot.Id,
@@ -701,11 +891,19 @@ func (v *VolumeSnapshotPortal) DeleteVolumeSnapshot() {
 		Metadata: snapshot.Metadata,
 		Context:  ctx.ToJson(),
 	}
-	if _, err = v.CtrClient.DeleteVolumeSnapshot(context.Background(), opt); err != nil {
-		log.Error("delete volume snapthot failed in controller service:", err)
-		return
-	}
 
+	if CONF.OsdsApiServer.InstallType == "thin" {
+		opt.DriverName = CONF.OsdsDock.EnabledBackends[0]
+		if _, err = v.DockClient.DeleteVolumeSnapshot(context.Background(), opt); err != nil {
+			log.Error("delete volume snapshot failed in controller service:", err)
+			return
+		}
+	} else {
+		if _, err = v.CtrClient.DeleteVolumeSnapshot(context.Background(), opt); err != nil {
+			log.Error("delete volume snapshot failed in controller service:", err)
+			return
+		}
+	}
 	v.Ctx.Output.SetStatus(StatusAccepted)
 	return
 }
